イベント&絞り込みの選択値を保持するように変更
